+++ src/it/scala/au/com/agl/arc/load/JDBCLoadSuite.scala
@@ -74,6 +74,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -115,6 +116,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -180,6 +182,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -245,6 +248,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -290,6 +294,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     try {
       connection = DriverManager.getConnection(sqlserverurl, connectionProperties)    
@@ -348,6 +353,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     try {
       connection = DriverManager.getConnection(sqlserverurl, connectionProperties)    
@@ -406,6 +412,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -492,6 +499,7 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false)
 
     val dataset = TestDataUtils.getKnownDataset
     dataset.createOrReplaceTempView(dbtable)
@@ -528,5 +536,72 @@ class JDBCLoadSuite extends FunSuite with BeforeAndAfter {
 
     assert(actual.first.getLong(0) == 2)
   } 
+
+  test("JDBCLoad: Structured Streaming") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit var arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=true)
+
+    val uuid = s"""a${UUID.randomUUID.toString.replace("-","")}"""
+
+    try {
+      connection = DriverManager.getConnection(postgresurl, connectionProperties)    
+      connection.createStatement.execute(s"""
+      DROP TABLE IF EXISTS sa.public.${uuid}
+      """)
+      connection.createStatement.execute(s"""
+      CREATE TABLE sa.public.${uuid} (
+        timestamp timestamp NULL,
+        value bigint NOT NULL
+      )
+      """)
+    } finally {
+      connection.close
+    }  
+
+    val readStream = spark
+      .readStream
+      .format("rate")
+      .option("rowsPerSecond", "1")
+      .load
+
+    readStream.createOrReplaceTempView(dbtable)    
+
+    load.JDBCLoad.load(
+      JDBCLoad(
+        name="dataset",
+        inputView=dbtable, 
+        jdbcURL=postgresurl, 
+        driver=DriverManager.getDriver(postgresurl),
+        tableName=s"sa.public.${uuid}", 
+        partitionBy=Nil, 
+        numPartitions=None, 
+        isolationLevel=None,
+        batchsize=None, 
+        truncate=None,
+        createTableOptions=None,
+        createTableColumnTypes=None,        
+        saveMode=Option(SaveMode.Overwrite), 
+        bulkload=Option(false),
+        tablock=None,
+        params=Map("user" -> user, "password" -> password)
+      )
+    )
+
+    Thread.sleep(2000)
+    spark.streams.active.foreach(streamingQuery => streamingQuery.stop)
+
+    val actual = { spark.read
+      .format("jdbc")
+      .option("url", postgresurl)
+      .option("user", user)
+      .option("password", password)      
+      .option("dbtable", s"(SELECT * FROM sa.public.${uuid}) result")
+      .load()
+    }
+
+    assert(actual.count > 0)
+  }  
 
 }